Mavenize pyRadiomics_GUI
