+++ pom.xml
@@ -1,0 +1,105 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+		http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.scijava</groupId>
+		<artifactId>pom-scijava</artifactId>
+		<version>19.2.0</version>
+		<relativePath />
+	</parent>
+
+	<groupId>com.github.salimkanoun</groupId>
+	<artifactId>pyRadiomics_GUI</artifactId>
+	<version>0.22.0-SNAPSHOT</version>
+
+	<name>pyRadiomics_GUI</name>
+	<description>JAVA GUI for pyRadiomics</description>
+	<url>N/A</url>
+	<inceptionYear>2017</inceptionYear>
+	<organization>
+		<name>N/A</name>
+		<url>N/A</url>
+	</organization>
+	<licenses>
+		<license>
+			<name>GNU General Public License v3.0</name>
+			<distribution>repo</distribution>
+		</license>
+	</licenses>
+
+	<developers>
+		<developer>
+			<id>salimkanoun</id>
+			<name>Salim Kanoun</name>
+			<url>http://imagej.net/User:Kanoun</url>
+			<roles>
+				<role>founder</role>
+				<role>lead</role>
+				<role>developer</role>
+				<role>debugger</role>
+				<role>reviewer</role>
+				<role>support</role>
+				<role>maintainer</role>
+			</roles>
+		</developer>
+	</developers>
+	<contributors>
+		<contributor>
+			<name>N/A</name>
+		</contributor>
+	</contributors>
+
+	<mailingLists>
+		<mailingList>
+			<name>ImageJ Forum</name>
+			<archive>http://forum.imagej.net/</archive>
+		</mailingList>
+	</mailingLists>
+
+	<scm>
+		<connection>scm:git:git://github.com/salimkanoun/pyRadiomics_GUI</connection>
+		<developerConnection>scm:git:git@github.com:salimkanoun/pyRadiomics_GUI</developerConnection>
+		<tag>HEAD</tag>
+		<url>https://github.com/salimkanoun/pyRadiomics_GUI</url>
+	</scm>
+	<issueManagement>
+		<system>GitHub Issues</system>
+		<url>https://github.com/salimkanoun/pyRadiomics_GUI/issues</url>
+	</issueManagement>
+	<ciManagement>
+		<system>None</system>
+	</ciManagement>
+
+	<properties>
+		<package-name>com.github.salimkanoun</package-name>
+		<license.licenseName>gpl_3</license.licenseName>
+		<license.copyrightOwners>Salim Kanoun</license.copyrightOwners>
+	</properties>
+
+	<repositories>
+		<repository>
+			<id>imagej.public</id>
+			<url>http://maven.imagej.net/content/groups/public</url>
+		</repository>
+	</repositories>
+
+	<dependencies>
+		<dependency>
+			<groupId>sc.fiji</groupId>
+			<artifactId>IO_</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-lang3</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.github.cliftonlabs</groupId>
+			<artifactId>json-simple</artifactId>
+			<version>2.3.1</version>
+		</dependency>
+	</dependencies>
+</project>
