+++ src/main/resources/logo/logo_pyRadiomics.png
