+++ src/main/resources/plugins.config
